+++ agents-flex-core/src/main/java/com/agentsflex/core/llm/functions/JavaNativeFunction.java
@@ -13,11 +13,11 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.agentsflex.core.functions;
+package com.agentsflex.core.llm.functions;
 
 import com.agentsflex.core.convert.ConvertService;
-import com.agentsflex.core.functions.annotation.FunctionDef;
-import com.agentsflex.core.functions.annotation.FunctionParam;
+import com.agentsflex.core.llm.functions.annotation.FunctionDef;
+import com.agentsflex.core.llm.functions.annotation.FunctionParam;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.reflect.InvocationTargetException;