+++ agents-flex-core/src/main/java/com/agentsflex/core/llm/functions/annotation/FunctionParam.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.agentsflex.core.functions.annotation;
+package com.agentsflex.core.llm.functions.annotation;
 
 import java.lang.annotation.*;