+++ agents-flex-llm/agents-flex-llm-chatglm/src/test/java/com/agentsflex/llm/chatglm/test/WeatherFunctions.java
@@ -1,7 +1,7 @@
 package com.agentsflex.llm.chatglm.test;
 
-import com.agentsflex.core.functions.annotation.FunctionDef;
-import com.agentsflex.core.functions.annotation.FunctionParam;
+import com.agentsflex.core.llm.functions.annotation.FunctionDef;
+import com.agentsflex.core.llm.functions.annotation.FunctionParam;
 import com.agentsflex.core.llm.client.HttpClient;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;